+++ R/runPDTMTAnalysis.R
@@ -83,6 +83,8 @@
 #'     interact with.
 #' @param complexFDRThr Numeric, FDR threshold for significance in testing
 #'     of complexes.
+#' @param maxNbrComplexesToPlot Numeric, the maximum number of significant
+#'     complexes for which to make separate volcano plots.
 #' @param seed Numeric, random seed to use for any non-deterministic
 #'     calculations.
 #' @param includeFeatureCollections Character vector, a subset of
@@ -118,7 +120,8 @@ runPDTMTAnalysis <- function(
     minlFC = 0, nperm = 250, volcanoAdjPvalThr = 0.05,
     volcanoLog2FCThr = 1, volcanoMaxFeatures = 25,
     volcanoS0 = 0.1, volcanoFeaturesToLabel = "",
-    addInteractiveVolcanos = FALSE, complexFDRThr = 0.1, seed = 42,
+    addInteractiveVolcanos = FALSE, complexFDRThr = 0.1,
+    maxNbrComplexesToPlot = 10, seed = 42,
     includeFeatureCollections, customComplexes = list(),
     complexSpecies = "all", complexDbPath
 ) {
@@ -163,7 +166,8 @@ runPDTMTAnalysis <- function(
         volcanoMaxFeatures = volcanoMaxFeatures,
         volcanoS0 = volcanoS0, volcanoFeaturesToLabel = volcanoFeaturesToLabel,
         addInteractiveVolcanos = addInteractiveVolcanos,
-        complexFDRThr = complexFDRThr, seed = seed,
+        complexFDRThr = complexFDRThr,
+        maxNbrComplexesToPlot = maxNbrComplexesToPlot, seed = seed,
         includeFeatureCollections = includeFeatureCollections,
         customComplexes = customComplexes, complexSpecies = complexSpecies,
         complexDbPath = complexDbPath)
@@ -192,7 +196,8 @@ runPDTMTAnalysis <- function(
              volcanoMaxFeatures = volcanoMaxFeatures,
              volcanoS0 = volcanoS0, volcanoFeaturesToLabel = volcanoFeaturesToLabel,
              addInteractiveVolcanos = addInteractiveVolcanos,
-             complexFDRThr = complexFDRThr, seed = seed,
+             complexFDRThr = complexFDRThr,
+             maxNbrComplexesToPlot = maxNbrComplexesToPlot, seed = seed,
              includeFeatureCollections = includeFeatureCollections,
              customComplexes = customComplexes, complexSpecies = complexSpecies,
              complexDbPath = complexDbPath)
@@ -236,9 +241,9 @@ runPDTMTAnalysis <- function(
                 "the QC pdf file: \n", paste(msng, collapse = "\n"))
     } else {
         pdf(sub("\\.Rmd$", "_PDTMTqc.pdf", outputFile), width = 14, height = 12)
-        print(plotPDTMTqc(pdOutputFolder = pdOutputFolder,
-                          pdResultName = pdResultName, masterOnly = FALSE,
-                          poiText = "", doPlot = TRUE, textSize = 4))
+        plotPDTMTqc(pdOutputFolder = pdOutputFolder,
+                    pdResultName = pdResultName, masterOnly = FALSE,
+                    poiText = "", doPlot = TRUE, textSize = 4)
         dev.off()
     }